Split out Base and Class for DynamicWriter/EventWriter
