+++ reflex.cabal
@@ -87,10 +87,12 @@ library
     Reflex,
     Reflex.Class,
     Reflex.Collection,
-    Reflex.EventWriter,
+    Reflex.EventWriter.Base,
+    Reflex.EventWriter.Class,
     Reflex.Dynamic,
     Reflex.Dynamic.Uniq,
-    Reflex.DynamicWriter,
+    Reflex.DynamicWriter.Base,
+    Reflex.DynamicWriter.Class,
     Reflex.FastWeak,
     Reflex.FunctorMaybe,
     Reflex.Host.Class,
